fix giftcard listall request dto

+++ colanta-backend/App/GiftCards/Application/Dtos/ListAllGiftCardsRequestDto.cs
@@ -9,14 +9,14 @@
 
     public class ListAllGiftCardsRequestCartDto
     {
-        public int grandTotal { get; set; }
+        public decimal grandTotal { get; set; }
         public object relationName { get; set; }
         public string redemptionCode { get; set; }
-        public int discounts { get; set; }
-        public int shipping { get; set; }
-        public int taxes { get; set; }
+        public decimal discounts { get; set; }
+        public decimal shipping { get; set; }
+        public decimal taxes { get; set; }
         public List<ListAllGiftCardsRequestItemDto> items { get; set; }
-        public int itemsTotal { get; set; }
+        public decimal itemsTotal { get; set; }
     }
 
     public class ListAllGiftCardsRequestClientDto
@@ -32,7 +32,7 @@
         public string id { get; set; }
         public string refId { get; set; }
         public string name { get; set; }
-        public int price { get; set; }
+        public decimal price { get; set; }
         public int quantity { get; set; }
     }
 }
